+++ MRMPService/MRPService/Classes/Modules/MRPPortal/Contracts/ManagementobjectStats.cs
@@ -50,7 +50,7 @@ namespace MRMPService.MRMPAPI.Contracts
         public long recovery_point_latency { get; set; }
 
         [JsonProperty("journal_size")]
-        public long journal_size { get; set; }
+        public double journal_size { get; set; }
         [JsonProperty("journal_used")]
         public double journal_used { get; set; }
     }
